requirements.txt
flask==3.0.0
python-dotenv==1.0.0
Flask-Script==2.0.6
flask-sqlalchemy==3.1.1
